sound: add Stream, ControlMessage and other types

+++ crates/sound/src/virtio_sound.rs
@@ -46,8 +46,8 @@ pub const VIRTIO_SND_S_IO_ERR: u32 = 0x8003;
 
 // device data flow directions
 
-pub const VIRTIO_SND_D_OUTPUT: u32 = 0;
-pub const VIRTIO_SND_D_INPUT: u32 = 1;
+pub const VIRTIO_SND_D_OUTPUT: u8 = 0;
+pub const VIRTIO_SND_D_INPUT: u8 = 1;
 
 // supported jack features
